Have handling mentality for var groups
THe group keyword is optional.  By default variables with the same
name may be joined on a plot.  This only make sense if variables are
redefined.

Sometimes different variables track different regions of a plot, but
these are plottable together to cover the most area.  The name lets
variables maintain thier identity, but the group keyword allows them
to be plotted together.

Short version:

  1. Variables of the same group occupy the same type of physical
     coordinates.

  2. The value of group defaults to the value of name.